code cleanup
- util.js for helper functions
- main.js for event handler registration
- documentation to functions added
- fix for video continuing to play after modal close
