+++ src/_includes/layouts/base.liquid
@@ -13,7 +13,7 @@
     {%- else %}
     <link rel="stylesheet" href="/css/index.css">
     {%- endif %}
-    <script src="/js/main.js"></script>
+    <script type="module" src="/js/main.js"></script>
   </head>
   <body>
     <div class="site-container">
